refactor: move layer handling to layer.ts
